fix build errors in migration
At this point I haven't done any manual or automated testing of the
backend. The interfaces for repositories have changed slightly so it
would probably be a good idea to check them. Esp in regards to
authenticated queries/mutations.

All of the nest plumbing isn't quite hooked up either to stand up the
full server but this seemed like a good stopping point.