Fix login/refresh token
Added a method to UserRepo to getById without requiring a token. I'm not
sure if this is really the best way to accomplish it but we're hacking a
bit here. This method should only be used in the refresh flow as it
doesn't check if you have access to that resource.